Constant off time converter working otherwise fine, but a HRTIM trap happened, need to rethink.
There is this super funny note on the manual:
"The external events used as triggers for reset, [...] are edge-sensitive even if EESNS
bit is reset (level-sensitive selection)"

So, the setting/resetting the output naturally works with level sensitive comparator event...
But starting the offtime counter is edge-sensitive, even when configured as level sensitive,
so it doesn't always start properly!

There doensn't seem to be simple way out. Need to rethink the control scheme.

Why, ST, why?