SBC SPI communication includes CRC8 now.
This simple addition was a two working day task, because the SPI peripheral is - surprise surprise -
still completely broken.

ST fixed an earlier silicon fault on their new design so that full peripheral resets through RCC
were not needed anymore - but they left some residual bugs behind, so that if CRC is enabled,
the peripheral STILL ends up in an internal fault state, if the RX DMA runs out of NDTR and RX
overruns (a completely normal scenario). Such fault state is hidden and not visible in any
SPI status register. The only way to get out of it seems to be - again - full reset through RCC.